refactor: use chainId from chains pkg
