+++ packages/swap-sdk/test/router.test.ts
@@ -2,7 +2,8 @@ import invariant from 'tiny-invariant'
 import { Percent, Token, CurrencyAmount } from '@pancakeswap/swap-sdk-core'
 import { Router } from '../src/router'
 import { Pair, Route, Trade, Native } from '../src/entities'
-import { ChainId, WNATIVE } from '../src/constants'
+import { WNATIVE } from '../src/constants'
+import { ChainId } from '../src'
 
 function checkDeadline(deadline: string[] | string): void {
   expect(typeof deadline).toBe('string')
