+++ packages/swap-sdk/test/trades.test.ts
@@ -1,7 +1,7 @@
 import { TradeType, Percent, Token, CurrencyAmount } from '@pancakeswap/swap-sdk-core'
 import { isTradeBetter } from '../src/trade'
 import { Pair, Route, Trade } from '../src/entities'
-import { ChainId } from '../src/constants'
+import { ChainId } from '../src'
 
 const BETTER_TRADE_LESS_HOPS_THRESHOLD = new Percent(50n, 10000n)
